refactor: move thread runtime buffer wiring
